this referred to the Issue-1
Actually, while running the 'mvn package' command, I was getting an error saying "No main method found." This happened because we skipped the repackaging part in the POM file inside the data module. Due to the older version style, it wasn't picking up. Now, I've added a separate file in the properties section of the data module's POM file, and the error is resolved. Even without a main method, it will execute. This refers to issue 1.